minor changes to UI and task attributes
Implemented the description and ID generation for tasks and included the
priority flag UI inside the TodoList componen. The implementation for
the on click outside for the flag button component is currently buggy,
will need to look into lifting the state of the TodoList up to the
parent or perhaps think about implementing global context state.